oh no it evals

+++ modules/system/boot/kernel.nix
@@ -4,7 +4,7 @@ let cfg = config.boot.kernel;
 in {
   options = {
     boot.kernel.enable = mkEnableOption (
-      "the FreeBSD kernel (sys). This can be disabled for jails where the host kernel is used")
+      "the kernel (sys). This can be disabled for jails where the host kernel is used")
       // {
         default = true;
       };
@@ -15,6 +15,38 @@ in {
       description = ''
         The package used for the kernel. This is just the derivation
         with the kernel and doesn't include out-of-tree modules.
+      '';
+    };
+
+    boot.kernel.isFreeBSD = mkOption {
+      default = cfg.package.meta.platforms == platforms.freebsd;
+      type = types.bool;
+      readOnly = true;
+      description = ''
+        Whether this system runs a FreeBSD kernel.
+      '';
+    };
+
+    boot.kernel.isOpenBSD = mkOption {
+      default = cfg.package.meta.platforms == platforms.openbsd;
+      type = types.bool;
+      readOnly = true;
+      description = ''
+        Whether this system runs an OpenBSD kernel.
+      '';
+    };
+
+    boot.kernel.flavor = mkOption {
+      default = if cfg.isFreeBSD then "freebsd"
+        else if cfg.isOpenBSD then "openbsd"
+        else throw "Couldn't detect known kernel; need either freebsd.sys or openbsd.sys";
+      type = types.str;
+      readOnly = true;
+      description = ''
+        A string indicating the kind of kernel being booted. One of:
+
+        - "freebsd"
+        - "openbsd"
       '';
     };
 
@@ -51,11 +83,19 @@ in {
   };
 
   config = mkIf cfg.enable (let
-    kernelPath = "${config.system.moduleEnvironment}/kernel/kernel";
-    modulePath = "${config.system.moduleEnvironment}/kernel";
+    inherit ({
+      freebsd = {
+        kernelPath = "${config.system.moduleEnvironment}/kernel/kernel";
+        modulePath = "${config.system.moduleEnvironment}/kernel";
+      };
+      openbsd = {
+        kernelPath = "${cfg.package}/bsd";
+        modulePath = "/not-supported";
+      };
+    }.${cfg.flavor}) kernelPath modulePath;
   in {
     system.build = { inherit (config.boot) kernel; };
-    system.moduleEnvironment = pkgs.buildEnv {
+    system.moduleEnvironment = mkIf cfg.isFreeBSD (pkgs.buildEnv {
       name = "sys-with-modules";
       paths = [ cfg.package ] ++ config.boot.extraModulePackages;
       pathsToLink = [ "/kernel" ];
@@ -63,7 +103,7 @@ in {
       postBuild = ''
         ${pkgs.buildPackages.freebsd.kldxref}/bin/kldxref $out/kernel
       '';
-    };
+    });
 
     system.systemBuilderCommands = ''
       if [ ! -f ${kernelPath} ]; then
@@ -73,10 +113,10 @@ in {
       fi
 
       ln -s ${kernelPath} $out/kernel
-      ln -s ${modulePath} $out/kernel-modules
+      test -e ${modulePath} && ln -s ${modulePath} $out/kernel-modules || true
     '';
 
-    boot.kernelEnvironment = {
+    boot.kernelEnvironment = mkIf cfg.isFreeBSD {
       module_path = modulePath;
       init_shell = config.environment.binsh;
     };
